GSYE-353: Move settlement and future market fixtures in their own module in tests
